Moved some code to cpp files.
